% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\name{add_eqv_aes}
\alias{add_eqv_aes}
\title{Add equivalent Americanized and base equivalent names to ggplot aesthetic
list}
\usage{
add_eqv_aes(aes_names)
}
\arguments{
\item{aes_names}{a character vector of aesthetic names}
}
\value{
a character vector of aesthetic names including any Americanized or
base R equivalent argument names accepted by ggplot2.
}
\description{
Add equivalent Americanized and base equivalent names to ggplot aesthetic
list
}
