% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGrid.R
\name{addGrid}
\alias{addGrid}
\title{Add a Grid to an Existing Plot.}
\usage{
addGrid(x = NULL, y = NULL, col = "lightgray", lwd = 1L, lty = 3L)
}
\arguments{
\item{x}{(integer, numeric) single integer specifies number of cells, numeric vector specifies vertical grid-lines}

\item{y}{(integer, numeric) single integer specifies number of cells, numeric vector specifies horizontal grid-lines}

\item{col}{(character) color of grid-lines}

\item{lwd}{(integer) line width of grid-lines}

\item{lty}{(integer) line type of grid-lines}
}
\description{
It is possible to use automatically determined grid lines (\code{x=NULL, y=NULL}) or specifying the number 
of cells \code{x=3, y=4} as done by \code{grid}. Additionally, x- and y-locations of grid-lines can be specified,
e.g. \code{x=1:10, y=seq(0,10,2)}.
}
\author{
Andre Schuetzenmeister \email{andre.schuetzenmeister@roche.com}
}
