% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addXlabTable.R
\name{addXlabTable}
\alias{addXlabTable}
\title{Create Tabular Environment from Matrix and Add it to Bottom Margin.}
\usage{
addXlabTable(mat, merge = TRUE, bmargin = 0.025, Label = list(),
  Text = list())
}
\arguments{
\item{mat}{(matrix) representing the table which will be added to lower margin;
rows represent factors, columns represent factor-levels, rownames will
be re-used as rownames of the table in the plot}

\item{merge}{(logial) TRUE = neighboring cells of 'mat' with the same content will
be merged to a wider cell}

\item{bmargin}{(numeric) value representing the height of the (b)ottom margin as proportion
of the total height of the figure region. Note, that this will determine the height
of the table in the bottom margin also depending on the amount of space available there.}

\item{Label}{(list) specifying all parameters applicable in function 'text', x- and y-values
will be set automatically for factor-labels (rows of the table)}

\item{Text}{(list) specifying all parameters applicable in function 'text', x- and y-values will
be set automatically for character strings appearing in the cells of the table}
}
\description{
This is a helper function intended to add tabular environments as X-axis
labels representing specific structure, e.g. formulas in boxplots.
Function takes a matrix and constructs a tabular environment from this.
}
\examples{
bp  <- BoxPlot(mpg~cyl:gear, mtcars, Xaxis=NULL)
mat <- getMatrix(mpg~cyl:gear, bp$names)
addXlabTable(mat)

# changing appearance for labels and cell-content

bp  <- BoxPlot(mpg~cyl:gear, mtcars, Xaxis=NULL)
addXlabTable(mat, Label=list(font=2, col="red"), Text=list(font=3, col="blue"))

}
\author{
Andre Schuetzenmeister \email{andre.schuetzenmeister@roche.com}
}
