% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\name{aes_arg_is_uneval}
\alias{aes_arg_is_uneval}
\title{Determine if an argument is an unevaluated expression}
\usage{
aes_arg_is_uneval(i)
}
\arguments{
\item{i}{the input object to be tests}
}
\description{
Determine if an argument is an unevaluated expression
}
