% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\name{allowed_aesthetics}
\alias{allowed_aesthetics}
\title{Get list of allowed aesthetics for a given geometry class}
\usage{
allowed_aesthetics(geom)
}
\arguments{
\item{geom}{a ggplot2 Geom object}
}
\value{
a character vector of the names of aesthetics permitted by this
geometry object (including Americanized names and base R equivalent names
also accepted by ggplot2)
}
\description{
Get list of allowed aesthetics for a given geometry class
}
