% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas-convenience-funcs.R
\name{append_split_terms}
\alias{append_split_terms}
\title{Wrap a formula in a helper list splitting response and predictor terms}
\usage{
append_split_terms(t)
}
\arguments{
\item{t}{the formula or \code{terms()} object to construct from}
}
\value{
a list with four elements, "terms" containing formula terms, "resp"
  containing a vector of reponse variable names, "covs" containing predictor
  covariate variable names and "vars" containing both response and covariate
  variable names.
}
\description{
Wrap a formula in a helper list splitting response and predictor terms
}
