% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-handling.r
\name{augment_predict}
\alias{augment_predict}
\title{Adds prediction outputs to data based on model and formula}
\usage{
augment_predict(.data, model, model.per = NULL, ...)
}
\arguments{
\item{.data}{data to use for model fitting}

\item{model}{model function to use (e.g `lm`)}

\item{model.per}{a formula representing all terms to group data by to fit
individual models (e.g. \code{~ am + vs})}

\item{...}{additional arguments to be passed to the model function, prefixed
with 'model' (eg. \code{model.singular.ok = FALSE})}
}
\value{
a data frame with new columns based on model fit
}
\description{
Adds prediction outputs to data based on model and formula
}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}
}
