% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenDoc.R
\name{AvailableDocs}
\alias{AvailableDocs}
\title{Show available documents}
\usage{
AvailableDocs()
}
\description{
Show available documents.
}
\author{
Kwame Okrah \email{okrah.kwame@gene.com}, Alexey Pronin \email{pronin.alexey@gene.com}
}
