% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmComp2.R
\name{bmComp2}
\alias{bmComp2}
\title{Given 2 biomarkers, find the contigency table and efficacy in the insection/unique portion}
\usage{
bmComp2(data, bm1 = NULL, bm1_pos_level = NULL, bm2 = NULL,
  bm2_pos_level = NULL, endpoint = NULL, trt.var = NULL,
  placebo.code = NULL, active.code = NULL, headerLevel = 2,
  armColors = c("black", "red"), ...)
}
\arguments{
\item{data}{input data; data frame only}

\item{bm1}{variable name for biomarker 1}

\item{bm1_pos_level}{the level of biomarker 1 which is considered positive. A character value will force bm1 into character.}

\item{bm2}{variable name for biomarker 2}

\item{bm2_pos_level}{the level of biomarker 2 which is considered positive}

\item{endpoint}{a vector of TTE endpoints}

\item{trt.var}{variable to indicate treatment variable; only 2 arms are allowed}

\item{placebo.code}{a character value to indicate the reference arm}

\item{active.code}{a character value to indicate the active arm.}

\item{headerLevel}{control the heading info for Rmd outoupt, default is 2}

\item{armColors}{color for each arm}

\item{...}{}
}
\value{
the original data frame with addition derived biomarker status variables
}
\description{
Given 2 biomarkers, find the contigency table and efficacy in the insection/unique portion
}
\examples{
```{r function,results='asis',warning=F,message=F,fig.width=7,fig.height=7}
b1 <- bmComp2(vad,'TC3IC3','TC3 or IC3','BEP263C3','>=50\%',c('PFSINV','OS'),trt.var='ARMCD1', placebo.code = 'B')

b2 <-  bmComp2(sample.data, 'KRAS.mutant', 'Mutant', 'CD8.ihc','1', 'OS', trt.var='Arm')
```
}
