% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas-convenience-funcs.R
\name{coerce_from_factor}
\alias{coerce_from_factor}
\title{Coerce dataframe variables to new type}
\usage{
coerce_from_factor(data, class_list, preserve = TRUE, suffix = ".old",
  verbose = FALSE)
}
\arguments{
\item{data}{the dataframe whose variables should be coerced to a new class}

\item{class_list}{a named list of classes to be coerced to from class factor,
with the list names being the variable names (e.g. \code{list(vs =
'character', carb = 'numeric')}). Classes are all coerced using the
function \code{as.<class>(levels(<var>)[<var>])}.}

\item{preserve}{whether original factor variables should be preserved}

\item{suffix}{a suffix to add to preserved variable names}

\item{verbose}{whether additional information should be printed to console}
}
\value{
the original dataframe with variables coerced to the target class and
  optionally preserved copies of the original variables.
}
\description{
Coerce many variables to new classes based on a named list of
  variable names with values of the character string of their target class.
}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}
}
