% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-handling.r
\name{copy_variable_attributes}
\alias{copy_variable_attributes}
\title{Copy attributes of dataframe variables from a source dataframe}
\usage{
copy_variable_attributes(obj, attr_src_obj)
}
\arguments{
\item{obj}{target dataframe}

\item{attr_src_obj}{source dataframe}
}
\value{
the target dataframe with attributes copied from the source dataframe
  for each variable.
}
\description{
particularly useful for copying variable label attributes, which
  are often not preserved during data transformation operations.
}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}
}
