% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePDF.R
\name{CreatePDF}
\alias{CreatePDF}
\title{Create pdf file}
\usage{
CreatePDF(pdfname, pdfparam)
}
\description{
Create pdf file
}
\examples{
\dontrun{
MainFunction <- function(..., pdfname = NULL, pdfparam = NULL) {
   # Do something here
   CreatePDF(pdfname, pdfparam)  
   # Do plotting here 
   CreatePDF(NULL) # Close the graphical device. 
   # Do something here
}

MainFunction(..., pdfname = "myplot.pdf", pdfparam = list(width=10, height=5))
}
}
\keyword{internal}

