% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoTabp.r
\name{demoTabp}
\alias{demoTabp}
\title{Title}
\usage{
demoTabp(ds, var, var.name = NULL, trt, varClass, varLevel, outF = NA, ...)
}
\arguments{
\item{ds}{input dataset}

\item{var}{the baseline variables to compare. do not include treatment variable here. #' @param var.name Optional, the variable text to display in the output. variable names supplied to \code{var.name} cannot overlap with its values (eg. a variable 'IC1' cannot contain values like 'IC1') in the output; in this case, use \code{var.name} to display a different text.}

\item{trt}{the variable name for the column head, trt variable must be included in varLevel unless trt is a factor variable already}

\item{varClass}{Optional, a named vector to specify the class for each var. eg varClass=c('age'='numeric','sex'='character', 'IC'='factor'). All variables that are specified as a factor will be converted into a factor before calling demoTab, and will be converted into an ordered factor before making pvalues. If \code{varClass} is not specified, the function will guess: if original numeric or integer -> numeric; if original character/factor and varLevels not specified -> char; if \code{varLevel} specified -> factor}

\item{varLevel}{a named vector to specify the factor levels.}

\item{outF}{a file path/name to save the results}

\item{...}{additional parameters to \code{fisher.test} or \code{cmh_test}}
}
\value{

}
\description{
Title
}
\examples{
demoTabp(ds = input, var=c('KRAS.mutant','KRAS.exprs'),  trt='Country', varClass=c( ), varLevel=list(  ),simulate.p.value = T)

}
