% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\name{filter_args}
\alias{filter_args}
\title{Helper function for ggpack to filter arguments based on a prefix}
\usage{
filter_args(prefix, args, sep = "\\\\.")
}
\arguments{
\item{prefix}{a string specifying the prefix to be used to filter args}

\item{args}{a list of arguments to be subset by the prefix match}

\item{sep}{a regex joining string between prefix and args.
Defaults to \code{"\\\\."}.}
}
\value{
a list of arguments that originally were prefaced by the
specified prefix, now with that prefix removed.
}
\description{
Helper function for ggpack to filter arguments based on a prefix
}
