% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\name{flatten_aesthetics_to_group}
\alias{flatten_aesthetics_to_group}
\title{Helper function to flatten specific aesthetics to group}
\usage{
flatten_aesthetics_to_group(mapping, ...)
}
\arguments{
\item{mapping}{a ggplot2 aesthetic mapping}

\item{...}{aesthetic handles to flatten into the group aesthetic as
interaciton terms}
}
\value{
the mapping aesthetics with specified aesthetics filtered out and
instead stored as interaction terms in the group aesthetics with any
preexisting group aesthetics
}
\description{
Helper function to flatten specific aesthetics to group
}
