% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestPlotBmrkEffects.R
\name{ForestPlotBmrkEffects}
\alias{ForestPlotBmrkEffects}
\title{Forest plot that examines biomarker effects within treatment arms}
\usage{
ForestPlotBmrkEffects(data, outcome.type = c("survival", "binary"),
  outcome.var, treatment.var, biomarker.var, covariate.var = NULL,
  strat.var = NULL, placebo.code = NULL, active.code = NULL,
  biomarker.code = NULL, alpha = 0.05,
  main = "Association of biomarker effect within treatment arms",
  sub = NULL, clip = NULL, cex.headings = 1.1, cex.note = 1, cols = 4,
  pdf.name = "Forestplot.pdf", pdf.param = list(width = 6, height = 4.5),
  par.param = list(cex = 1.2, cex.main = 1.5, cex.sub = 1, cex.axis = 1))
}
\arguments{
\item{data}{input data frame. Rows are patients and columns are variables (e.g. demographics variables, time to event variables, 
biomarker variables, treatment indicator, etc.). One patient per row.}

\item{outcome.type}{type of the outcome variable. Default is \code{c("survival", "binary")}}

\item{outcome.var}{name of the outcome varible. For time to event outcome, both names of the 'time to event' and 'censorship' variables need to be specified.}

\item{treatment.var}{name of the treatment variable}

\item{biomarker.var}{name of the biomarker variable}

\item{covariate.var}{a vector specifying the covariate variables to be adjusted in the model. Default is set to NULL, meaning no adjustment.}

\item{strat.var}{name of the stratification variables. Default is set to NULL, meaning no stratification.}

\item{placebo.code}{name of the control arm of the treatment variable}

\item{active.code}{of the treatment/experimental arm of the treatment variable}

\item{biomarker.code}{levels of the biomarker variable}

\item{alpha}{type I error rate. Default is 0.05.}

\item{main}{main title of the forest plot. Default is "Association of biomarker effect within treatment arms".}

\item{sub}{footnote under the forest plot. Default is NULL.}

\item{clip}{range of the x-axis of the forest plot. Default is NULL.}

\item{cex.headings}{amount of magnification of headings of the forest plot relative to cex. Default is 1.1.}

\item{cex.note}{amount of magnification of the note. Default is 1.}

\item{cols}{Color of the 'effect size' displayed in the forest plot.}

\item{pdf.name}{name of output pdf file. If it's NULL, the plots will be displayed but not saved as pdf. Default is "Forestplot.pdf".}

\item{pdf.param}{a list of parameters that define pdf graphics device. See \code{\link{pdf}}. Default is \code{list(width=6, height=4.5)}.}

\item{par.param}{a list of parameters that define graphcial parameters. See \code{\link{par}}. Default is \code{list(mar=c(4,4,3,2))}.}
}
\description{
This function creates a forest plot along with table with summary statistics to examine the biomarker effects within treatment arms.
}
\author{
Alexey Pronin \email{pronin.alexey@gene.com}, Ning Leng \email{leng.ning@gene.com}, and previous team members (see DESCRIPTION)
}

