% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-geom-text-table.r
\docType{data}
\name{GeomTextTable}
\alias{GeomTextTable}
\title{A ggplot2 Geom object to add a table of text values to a plot}
\format{An object of class \code{GeomTextTable} (inherits from \code{Geom}, \code{ggproto}) of length 7.}
\usage{
GeomTextTable
}
\description{
A ggplot2 Geom object to add a table of text values to a plot
}
\keyword{datasets}
