% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.colors.R
\name{get.colors}
\alias{get.colors}
\title{Choose an R standard color or get the RGB-code of an arbitrary color.}
\usage{
get.colors(continuous = TRUE)
}
\arguments{
\item{continuous}{(logical) TRUE = one or multiple colors can be chosen from the RGB color-space via a tcltk-widget
FALSE = a grid is drawn where R standard colors can be selected by clicking on colored-tiles}
}
\description{
Function lets the user choose a color via a tcltk widget in case it is called with default value of
paramater 'continuous'. Otherwise, it plots a grid of differently colored tiles, which can be used to select as many
colors as desired by simply clicking at the corresponding tiles.
}
\examples{
## get R standard color
 get.colors(FALSE)

## get RGB-code (hexadecimal) via tcltk-widget
 \dontrun{
     get.colors()
 }
 
}
\author{
Andre Schuetzenmeister \email{andre.schuetzenmeister@roche.com}
}
