% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas-convenience-funcs.R
\name{get_lsmeans_specs_predictors}
\alias{get_lsmeans_specs_predictors}
\title{Return predictors from specs input}
\usage{
get_lsmeans_specs_predictors(specs)
}
\arguments{
\item{specs}{a specs parameter to be passed to \code{lsmeans::lsmeans()}}
}
\value{
a character vector of the variables from the specs object
}
\description{
\code{lsmeans::lsmeans()} accepts a bunch of different variable
  types to its \code{specs} parameter. This function handles that variety of
  variable types to figure out what variables were specified.
}
