% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatrix.R
\name{getMatrix}
\alias{getMatrix}
\title{Construct Matrix from Formula and Names.}
\usage{
getMatrix(form, labels, split="\\\\.")
}
\arguments{
\item{form}{(formula) object from which the matrix should be constructed}

\item{labels}{(character) the 'names" element of the list returned by e.g.
\code{\link{BoxPlot}} or \code{\link{boxplot}}}

\item{split}{(character) symbol or string specifying the split symbol or
string separating factor-levels in 'labels'}
}
\description{
Function is a helper function for adding tables as X-axis labels
representing the structure of crossed (nested) factor-levels.
}
\examples{
bp  <- boxplot(mpg~cyl:gear, mtcars, plot=FALSE)
mat <- getMatrix(mpg~cyl:gear, bp$names)
mat

}
\author{
Andre Schuetzenmeister \email{andre.schuetzenmeister@roche.com}
}
