% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpkg-.r
\name{ggbundle_filter_args}
\alias{ggbundle_filter_args}
\title{Helper function for ggbundle to filter arguments based on a prefix}
\usage{
ggbundle_filter_args(prefix, args)
}
\description{
Helper function for ggbundle to filter arguments based on a prefix
}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}
}
