% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\name{ggpack_aes_from_dots}
\alias{ggpack_aes_from_dots}
\title{Helper function to cull aesthetic mapping from parameter inputs in ellipses}
\usage{
ggpack_aes_from_dots(...)
}
\arguments{
\item{...}{ellipses containing aesthetics mapping}
}
\value{
the ellipses arguments passed through a ggplot2::aes() call after
filtering for aesthetics in ggplot2:::.all_aesthetics()
}
\description{
Helper function to cull aesthetic mapping from parameter inputs in ellipses
}

