% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\name{ggpack_filter_aesthetics}
\alias{ggpack_filter_aesthetics}
\title{Helper function to filter aesthetic mappings based on geometry}
\usage{
ggpack_filter_aesthetics(geom, mapping)
}
\arguments{
\item{geom}{a ggplot2 Geom object}

\item{mapping}{a ggplot2 aesthetic mapping}
}
\value{
the mapping filtered by accepted aesthetics for the given Geom
}
\description{
Helper function to filter aesthetic mappings based on geometry
}
