% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\name{ggpack_filter_args}
\alias{ggpack_filter_args}
\title{Helper function for ggpack to filter arguments based on a prefix}
\usage{
ggpack_filter_args(prefix, args)
}
\description{
Helper function for ggpack to filter arguments based on a prefix
}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}
}
