% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\docType{class}
\name{ggpacket-class}
\alias{ggpacket}
\alias{ggpacket-class}
\title{A simple S4 class to contain wrapped ggplots}
\description{
A simple S4 class to contain wrapped ggplots
}
\note{
eventually this class could be used to handle pass
through of geom-specific parameters. For now it just collects
a list with cleaner syntax for adding ggproto objects.
}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gnee.com}
}

