% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\name{ggpk_decorators}
\alias{ggpk_decorators}
\title{Wrapper for common decorators to package}
\usage{
ggpk_decorators(...)
}
\arguments{
\item{...}{arguments to be passed to any of the bundled
decorator plot functions, blank, xlab, ylab, labs or theme
prefaced by any of those strings ('labs.title').}
}
\value{
a ggpacket object containing a collection of
decorating plot functions.
}
\description{
Wrapper for common decorators to package
}
