% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpk-stat-ribbons.r
\name{ggpk_stat_ribbon}
\alias{ggpk_stat_ribbon}
\title{ggplot wrapper for plotting ribbon, taking a stats string or function set
as an argument}
\usage{
ggpk_stat_ribbon(mapping = NULL, data = NULL, show.counts = FALSE,
  fun.data = "mean_se", fun.args = list(), ...)
}
\arguments{
\item{mapping}{a ggplot aesthetic mapping}

\item{data}{data to pass to ggplot2::stat_summary functions}

\item{show.counts}{True, False, "label" or "table" to display either as
labels at each datapoint or as a table at the top of the plot}
}
\value{
a ggpacket object that can be added to any ggplot to
draw a collection of layers
}
\description{
ggplot wrapper for plotting ribbon, taking a stats string or function set
as an argument
}
\note{
When specifying arguments for wrapped ggplot geometries, prefix
the parameter with geometry specifiers: "ribbon" for ribbon geometry
"line" for central line geometry, or "label" for count labels. e.g.
\code{ggpk_stat_ribbon(ribbon.alpha = 0.8, line.alpha = 0.6)}
}
\examples{
library(tidyverse) # for dplyr, ggplot
library(lubridate)

nasa \%>\% as_tibble \%>\%
  mutate(date = ymd("0000/01/01") + months(month) + years(year)) \%>\%
  mutate(hemisphere = ifelse(lat>0,"Northern", "Southern")) \%>\%
  mutate(temperature = temperature - 273) \%>\%

  # plotting
  ggplot() +
    aes(x=date, y=temperature) +
    ggpk_stat_ribbon(fun.data = 'deciles') +
    facet_grid(hemisphere ~ .) +
    labs(title="Temperature by Hemisphere")

}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}
}

