% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggwrapper-.r
\name{ggwrap_filter_args}
\alias{ggwrap_filter_args}
\title{Helper function for ggwrap to filter arguments based on a prefix}
\usage{
ggwrap_filter_args(prefix, args)
}
\description{
Helper function for ggwrap to filter arguments based on a prefix
}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}
}

