% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\docType{methods}
\name{initialize,ggpacket-method}
\alias{initialize,ggpacket-method}
\alias{initialize,ggpacket-method}
\title{Initialize empty ggpacket}
\usage{
\S4method{initialize}{ggpacket}(.Object, ggproto_obj = NULL, label = NULL)

\S4method{initialize}{ggpacket}(.Object, ggproto_obj = NULL, label = NULL)
}
\arguments{
\item{.Object}{ggpacket object to be initialized}

\item{ggproto_obj}{an object of class ggproto to add to the ggpacket's internal list}

\item{label}{an optional label to be able to index this component within the ggpacket}
}
\description{
Initialize empty ggpacket

Initialize new object by adding ggproto to list with name label
}
