% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpkg-.r
\docType{methods}
\name{initialize,ggpkg-method}
\alias{initialize,ggpkg-method}
\title{Initialize new object by adding ggproto to list with name label}
\usage{
\S4method{initialize}{ggpkg}(.Object, ggproto_obj = NULL, label = NULL)
}
\description{
Initialize new object by adding ggproto to list with name label
}
