% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggwrapper-.r
\docType{methods}
\name{initialize,ggwrapper-method}
\alias{initialize,ggwrapper-method}
\title{Initialize new object by adding ggproto to list with name label}
\usage{
\S4method{initialize}{ggwrapper}(.Object, ggproto_obj = NULL, label = NULL)
}
\description{
Initialize new object by adding ggproto to list with name label
}

