% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input}
\alias{input}
\title{Data example}
\format{A data frame with 550 rows and 18 variables.}
\usage{
data(input)
input
}
\description{
Data sample from a study.
}
\author{
Alexey Pronin \email{pronin.alexey@gene.com}, Ning Leng \email{leng.ning@gene.com}
}
