% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.p.r
\name{inter.p}
\alias{inter.p}
\title{Title likelihood ratio test of interaction effect}
\usage{
inter.p(
  ds,
  tte = "Surv(time = tte, event=event)",
  reducedText = "BM + trt",
  fullText = "BM * trt"
)
}
\arguments{
\item{ds}{input data set}

\item{tte}{a text string that returns a Survival Object using \code{Surv}}

\item{reducedText}{main effect model without interaction term}

\item{fullText}{full model with interaction term}
}
\value{

}
\description{
Title likelihood ratio test of interaction effect
}
\examples{
data(input); p1 <- inter.p(ds=subset(input, !is.na(KRAS.mutant)), tte='Surv(time = PFS, event=PFS.event)', reducedText = 'KRAS.mutant + Arm', fullText = 'KRAS.mutant * Arm')

}
