% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longbmkr.R
\docType{data}
\name{longbmkr}
\alias{longbmkr}
\title{Simulated longitudinal biomarker data}
\format{A data frame with 8,500 rows and 8 variables:
\describe{
  \item{pid}{Patient Identifier}
  \item{trt}{Treatment Arm (1, 0)}
  \item{sex}{Patient Sex (m, f)}
  \item{age}{Patient Age}
  \item{edu}{Patient years of education}
  \item{bmkr}{Baseline biomarker reading}
  \item{vm}{Patient visit time in months}
  \item{ep}{Biomarker endpoint reading}
}}
\usage{
longbmkr
}
\description{
Contains the following variables
}
\keyword{datasets}
