% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-handling.r
\name{na_if_at}
\alias{na_if_at}
\title{Filter only specific columns, leaving the rest of the dataframe intact}
\usage{
na_if_at(.tbl, .cols, ...)
}
\arguments{
\item{.tbl}{a dataframe or tibble to be sent to mutate}

\item{.cols}{a dplyr::vars() column selection similar to `dplyr::select()` or 
numeric indices of columns to select}

\item{...}{filtering functions to filter column data}
}
\value{
a dataframe with the specified columns filtered by the specified conditions
}
\description{
Useful for filtering down to a single datapoint in long data formats, for
instance when selecting baseline values when visitation date is a specific
code.
}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}
}

