% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas-convenience-funcs.R
\name{numeric_means}
\alias{numeric_means}
\title{List means over numeric variables}
\usage{
numeric_means(data, potential_vars = names(data), verbose = FALSE)
}
\arguments{
\item{data}{A dataframe upon which means will be calculated}

\item{potential_vars}{A character vector of potential variable names to
consider}

\item{verbose}{include verbose output of numeric means calculations}
}
\value{
A named list with names as the numeric subset of the variable names
  specified in potential_vars and values of the means
}
\description{
List means over numeric variables
}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}
}
