% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenDoc.R
\name{OpenDoc}
\alias{OpenDoc}
\title{Access package documents}
\usage{
OpenDoc(doc, verbose = FALSE)
}
\arguments{
\item{doc}{a character vector specifying the document to open;
see \code{\link{AvailableDocs}}}

\item{verbose}{a logical vector indicating whether the path of the
document should be printed (Default = FALSE)}
}
\description{
This function allows the user the get access to pdf documents associated
with this package.
}
\author{
Kwame Okrah, \email{okrah.kwame@gene.com}, Alexey Pronin \email{pronin.alexey@gene.com}
}
