% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDens.R
\name{PlotDens}
\alias{PlotDens}
\title{Plot Density}
\usage{
PlotDens(data, var, log2 = FALSE, col = "blue", text.font = 3,
  main = paste("Distribution of", var), add.num = 0, xlab = paste(var,
  ifelse(log2 == TRUE, "(log2 scale)", "")), pdf.name = NULL,
  pdf.param = list(width = 6, height = 4.5), par.param = list(mar = c(4, 4,
  3, 2)))
}
\arguments{
\item{data}{Input data frame. Rows are patients and columns are variables (e.g. demographics variables, time to event variables, 
biomarker variables, treatment indicator, etc.). One patient per row.}

\item{var}{Name of a variable (e.g. a biomarker variable). Should be in colnames of \code{data}.}

\item{log2}{If TRUE, computes binary (i.e. base 2) logarithm. Default is FALSE.}

\item{col}{The color of the line segments. Default is blue.}

\item{text.font}{Legend text font size. Default is 3.}

\item{main}{The main title. Default is \code{paste("Distribution of", var)}. A variable name will be added at the end of the title.}

\item{add.num}{The constant to add to all values. Helps to avoid applying log transformation on 0 or negative values. Default is 0.}

\item{xlab}{X axis label. Default is \code{paste(var, ifelse(log2 == TRUE, "(log2 scale)", ""))}. A variable name will be added at the beginning of the label.}

\item{pdf.name}{Name of output pdf file. If it's NULL (default), the plots will be displayed but not saved as pdf.}

\item{pdf.param}{A list of parameters that define pdf graphics device. See \code{\link{pdf}}. Default is \code{list(width=6, height=4.5)}.}

\item{par.param}{A list of parameters that define graphcial parameters. See \code{\link{par}}. Default is \code{list(mar=c(4,4,3,2))}.}
}
\value{
Histogram and density will be shown. Summary statistics will be also shown on the plots.
}
\description{
Generate density plot for a continuous covariate.
}
\examples{
data(input)
PlotDens(data=input, var="KRAS.exprs", log2=TRUE)

}
\author{
Alexey Pronin \email{pronin.alexey@gene.com}, Ning Leng \email{leng.ning@gene.com}, and previous team members (see DESCRIPTION)
}
