% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotKM.R
\name{PlotKM}
\alias{PlotKM}
\title{Generate KM curve(s) for full population or subgroups defined by single factor}
\usage{
PlotKM(data, tte, cens, subgroups = NULL, subgroups.levels = NULL,
  subgroups.labels = NULL, bep = NULL, bep.indicator = 1,
  plot.nrisk = TRUE, nrisk.interval = 2, cex.nrisk = 0.8,
  plot.grid = TRUE, grids = seq(0, 1, 0.1), plot.legend = TRUE,
  col = NULL, lty = 1, lwd = 3, xlab = "Months To Event Or Censoring",
  ylim = c(0, 1), xlim = NULL, ylab = "Survival Probability", main = "",
  sub = "", plot.median = FALSE, median.cex = 0.8, digits = 2,
  pdf.name = NULL, pdf.param = list(height = 5), par.param = list(mar =
  c(12, 9, 3, 2)))
}
\arguments{
\item{data}{Input data frame. Rows are patients and columns are variables (e.g. demographics variables, time to event variables, 
biomarker variables, treatment indicator, etc.). One patient per row.}

\item{tte}{column name that indicates the time to event variable}

\item{cens}{column name that indicates the censoring variable associated to tte}

\item{subgroups}{name of the column which indicates the subgroups (e.g. treatment group).
this column is expected to be categorical. If this column is in character class and subgroups.levels is not specified, 
it will be converted to a factor by factor() function. If subgroups.levels is defined, the column will be converted to
a factor following the level order in subgroups.levels. 
In the legend, the subgroups will be ordered based on the order of factor levels.}

\item{subgroups.levels}{levels in the subgroups}

\item{subgroups.labels}{preferred labels for the subgroups. The order in subgroups.labels should match subgroups.levels}

\item{bep}{name of the column which indicates subpopulation (e.g. biomarker evaluable population)
If parameter bep is not defined, the KM curve(s) will be draw using all samples.
If bep is defined, the KM curve(s) will be draw using only samples in BEP.}

\item{bep.indicator}{In the subpopulation column, which value is used
to define the subpopulation (e.g. biomarker evaluable population). 
Default is 1. The non-subpopulation enrties is not allowed to be specified as NA.}

\item{plot.nrisk}{whether show number of patients at risk at the below the graph. If it is specified as TRUE, number of patients
at risk will be summarized by subgroup.}

\item{nrisk.interval}{interval to summarize number of patients at risk . Default is to summarize every 2 (months)}

\item{cex.nrisk}{font size for the number of patients at risk.}

\item{plot.grid}{whether show horizontal grids}

\item{grids}{horizontal grids}

\item{plot.legend}{whether show legend}

\item{xlim, ylab, xlab, main, col, lty, lwd, sub, ylim}{see \code{\link{plot}}}

\item{plot.median}{whether show median TTE of each subgroup. (won't show if median TTE is NA)}

\item{median.cex}{font size of marked median. This parameter will be ignored if plot.median=FALSE}

\item{digits}{digits for rounding}

\item{pdf.name}{name of output pdf file. If it is NULL (default), the plots will be displayed but not saved as pdf}

\item{pdf.param}{A list of parameters that define pdf graphics device. See \code{\link{pdf}}. 
for example, pdf.param=list(width=10, height=5)}

\item{par.param}{A list of parameters that define graphcial parameters. See \code{\link{par}}.
for example, par.param=list(mfrow=c(3,1), mar=c(4,4,3,2))}

\item{...}{additional parameters for \code{\link{plot}}}
}
\description{
Generate KM curve(s) for full population or subgroups defined by single factor
}
\note{
This function generates KM curve(s) for full population (when parameter subgroups is not defined) 
or subgroups (when parameter subgroups is defined).
}
\examples{
data(input)
sample.data <- input
PlotKM(data=sample.data, tte="OS",cen="OS.CNSR", main="OS ITT by treatment", subgroups="Arm")
}
\author{
Ning Leng \email{leng.ning@gene.com}, Alexey Pronin \email{pronin.alexey@gene.com}, and previous team members (see DESCRIPTION)
}

