% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLong.r
\name{PlotLong}
\alias{PlotLong}
\title{Wrapper for fitting model and adjusting to predicted values and passing newly
fitted model to ggplot geom_stat_ribbon}
\usage{
PlotLong(data, mapping = NULL, model = lm, model.per = NULL,
  model.formula = NULL, facet.fun = NULL, plot.style = "ribbons", ...)
}
\arguments{
\item{data}{data to use for model fitting and plotting}

\item{mapping}{ggplot aesthetic mapping to use for plotting}

\item{model}{model function to use for fitting; defaults to lm}

\item{model.per}{grouping variables by which to isolate individual models}

\item{model.formula}{model formula to use for model fitting}

\item{facet.fun}{function to use for ggplot faceting in ggplot2::facet_grid}

\item{plot.style}{one of 'errorbars' or 'ribbons'}

\item{...}{Additional arguments allow for an assortion of additional
functionality.
\itemize{
\item{"Aesthetics"}{
First, arguments which can be used as any of ggplot's
default aesthetics will be pulled from \code{...} args.
}.
\item{"ggplot layers"}{
Arguments are passed to underlying ggplot components by prefixing value
with ggplot call name. ("ribbons", "line", "text", "facet", "xlab", "ylab",
"labs" or "theme" - e.g. \code{ribbons.color = 'red')}).
}
\item{"Model fitting"}{
Arguments with prefixes of "model" will be passed to the model function if
one is provided.
}
\item{"ggpackets"}{
Additional arguments passed to
\code{\link[gClinBiomarker]{ggpk_ribbons}} or
\code{\link[gClinBiomarker]{ggpk_line_errorbar}}.
}
}}
}
\value{
a ggplot object
}
\description{
Wrapper for fitting model and adjusting to predicted values and passing newly
fitted model to ggplot geom_stat_ribbon
}
\examples{
# load data
nasa.data <- as.data.frame(dplyr::nasa)

# default representation
PlotLong(nasa.data, x = month, y = temperature)

# a more appropriate representation for large n
PlotLong(nasa.data, x = month, y = temperature, fun.data = 'quartiles')

# include linear adjustment accounting for amount of ozone
PlotLong(nasa.data, x = month, y = temperature,
         formula = temperature ~ ozone, fun.data = 'deciles',
         show.counts = T)

# adjusting by independent models for the northern and southern hemispheres
library(dplyr) # needed for \%>\%
PlotLong(nasa.data \%>\% mutate(hemi=ifelse(lat>0, "North", "South")),
         x = month, y = temperature, formula = temperature ~ ozone,
         model.per = ~ hemi, facet.fun = ~ hemi, fun.data = 'deciles',
         xlab = "Month", ylab = "Temperature Adjusted for Ozone",
         labs.title = "Temperature by Hemisphere",
         labs.caption = "*idependent models fit per hemisphere")

# including a table of value counts and subsetting value data to specific
# months
library(dplyr) # needed for \%>\%
PlotLong(nasa.data \%>\% mutate(hemisphere=ifelse(lat > 0, "North", "South")),
         x = month, y = temperature, group = hemisphere,
         color = hemisphere, fill = hemisphere,
         formula = temperature ~ ozone,
         model.per = ~ hemisphere, fun.data = 'deciles',
         plot.style = 'errorbars',
         show.counts = 'table',
         label.data = . \%>\% filter(month \%in\% c(1, 6, 12)),
         label.hjust = 'inward',
         xlab = "Month", ylab = "Temperature Adjusted for Ozone",
         labs.title = "Temperature by Hemisphere",
         labs.caption = "*idependent models fit per hemisphere")

}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}
}
