% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePDF.R
\name{PlotParam}
\alias{PlotParam}
\title{Set graphical parameters}
\usage{
PlotParam(pdf.name, pdf.param, par.param)
}
\arguments{
\item{pdf.name}{A name of a pdf file.}

\item{pdf.param}{A list of parameters that define pdf graphics device. See \code{\link{pdf}}.}

\item{par.param}{A list of parameters that define graphcial parameters. See \code{\link{par}}.}
}
\description{
\code{PlotParam} can be used to set graphical parameters for a pdf file 
or the default screen device.
}
\note{
The \code{PlotParam} function is designed to restore graphical parameters \code{par} 
to the default values after execution.
}
\examples{
\dontrun{
plotKMsub <- function(..., 
                      pdf.name = NULL, 
                      pdf.param = NULL, 
                      par.param = NULL) {
   # Do something here.
   PlotParam(pdf.name, pdf.param, par.param)  
   # Do plotting here. 
   PlotParam() # Close the pdf graphical device if it was open and 
               # set graphical parameters to the default values. 
   # Do something here.
}

# Define both pdf graphics device and graphcial parameters.
plotKMsub(..., pdf.name = "myplot.pdf", 
               pdf.param = list(width=10, height=5), 
               par.param = list(mfrow=c(2, 2)))

# Define graphcial parameters. A pdf file is not created.
# The graph is pdoduced in the default screen device.
plotKMsub(..., par.param = list(mar=c(7,7,7,7)))   
            
}
}
\author{
Alexey Pronin, Ning Leng
}

