% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\docType{methods}
\name{+,gg,ggpacket-method}
\alias{+,gg,ggpacket-method}
\title{Combine ggpacket contents with gg call to incorporate seamlessly with ggplot calls}
\usage{
\S4method{+}{gg,ggpacket}(e1, e2)
}
\arguments{
\item{e1}{ggplot gg class object to add to ggpacket internal list}

\item{e2}{ggpacket object}
}
\description{
Combine ggpacket contents with gg call to incorporate seamlessly with ggplot calls
}
