% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpkg-.r
\docType{methods}
\name{+,ggpkg,ANY-method}
\alias{+,ggpkg,ANY-method}
\title{Primitive methods for adding ggpkgs to various ggplot objects}
\usage{
\S4method{+}{ggpkg,ANY}(e1, e2)
}
\description{
Primitive methods for adding ggpkgs to various ggplot objects
}
