% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggwrapper-.r
\docType{methods}
\name{+,ggwrapper,ANY-method}
\alias{+,ggwrapper,ANY-method}
\title{Primitive methods for adding ggwrappers to various ggplot objects}
\usage{
\S4method{+}{ggwrapper,ANY}(e1, e2)
}
\description{
Primitive methods for adding ggwrappers to various ggplot objects
}

