% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\docType{methods}
\name{+,NULL,ggpacket-method}
\alias{+,NULL,ggpacket-method}
\title{Add NULL to a ggpacket to return the ggpacket}
\usage{
\S4method{+}{`NULL`,ggpacket}(e1, e2)
}
\arguments{
\item{e1}{NULL value}

\item{e2}{the ggpacket object}
}
\value{
when NULL is added to a ggpacket object, the original object is returned
}
\description{
Add NULL to a ggpacket to return the ggpacket
}
