% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-io.r
\name{print_to_string}
\alias{print_to_string}
\title{Capture output and return as string}
\usage{
print_to_string(expr, condense = FALSE)
}
\arguments{
\item{expr}{the expression to be evaluated}

\item{condense}{whether empty lines should be stripped from the output}
}
\value{
a single string of the captured output
}
\description{
Take output from a segment of code and return it as a single
  string.
}
