% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadVAD.R
\name{ReadVAD}
\alias{ReadVAD}
\title{Read a VAD}
\usage{
ReadVAD(file, write.csv.to = NULL)
}
\arguments{
\item{file}{A SAS dataset.}

\item{write.csv.to}{Defines where both data and data.specs csv files will be written to. 
The csv files will be written in the form name_date.csv and name_specs_date.csv.
Default is NULL.}
}
\value{
A list with two data frames: data and data.specs.
}
\description{
Read a VAD and create specs for it.
}
\note{
The ReadVAD() function reads in a VAD and creates data specs.
}
\examples{
\dontrun{
ReadVAD("ars.sas7bdat")  
ReadVAD("ars.sas7bdat", ".")  
ReadVAD("ars.sas7bdat", "Desktop/R")
ReadVAD("Desktop/SAS/ars.sas7bdat", "Desktop/R")
}

}
\author{
Alexey Pronin \email{pronin.alexey@gene.com}, Ning Leng \email{leng.ning@gene.com}
}
