% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\name{remove_aesthetics}
\alias{remove_aesthetics}
\title{Helper function to filter out aesthetics from mappings}
\usage{
remove_aesthetics(mapping, ...)
}
\arguments{
\item{mapping}{aesthetic mapping to use for filtering}

\item{...}{mapping labels to filter out}
}
\value{
an aesthetic mapping with filtered mappings removed
and group mapping set as interaction terms of all non axial
terms.
}
\description{
Helper function to filter out aesthetics from mappings
}
