% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RspBarPlot.R
\name{RspBarPlot}
\alias{RspBarPlot}
\title{Description 1}
\usage{
RspBarPlot(outcome.var, reference.var, ref.popname = "All",
  treatment.var = NULL, biomarker.var = NULL, biomarker.varname = NULL,
  biomarker.code = NULL,
  main = "Association of biomarker expression with response rate",
  sub = NULL, return.table = TRUE, dichotomized = TRUE,
  pdf.name = "rsp.barplot.pdf", pdf.param = list(width = 6, height = 4.5),
  par.param = list(mar = c(4, 4, 3, 2)))
}
\arguments{
\item{outcome.var}{vector specifying the outcome variable.}

\item{reference.var}{???.}

\item{ref.popname}{???. Default is "All".}

\item{treatment.var}{name of the treatment variable. Default is NULL.}

\item{biomarker.var}{name of the biomarker variable. Default is NULL.}

\item{biomarker.varname}{???. Default is NULL.}

\item{biomarker.code}{???. Default is NULL.}

\item{main}{main title of the barplot. Default is "Association of biomarker expression with response rate".}

\item{sub}{footnote under the bar plot. Default is NULL.}

\item{return.table}{???. Default is TRUE.}

\item{dichotomized}{???. Default is TRUE.}

\item{pdf.name}{name of output pdf file. If it's NULL, the plots will be displayed but not saved as pdf. Default is "rsp.barplot.pdf".}

\item{pdf.param}{list of parameters that define pdf graphics device. See \code{\link{pdf}}. Default is \code{list(width=6, height=4.5)}.}

\item{par.param}{list of parameters that define graphcial parameters. See \code{\link{par}}. Default is \code{list(mar=c(4,4,3,2))}.}
}
\value{
???
}
\description{
Description 2
}
\examples{
data(input)

}
\author{
Alexey Pronin \email{pronin.alexey@gene.com}, Ning Leng \email{leng.ning@gene.com}, and previous team members (see DESCRIPTION)
}
