% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-.r
\name{safe_private_export}
\alias{safe_private_export}
\title{A function for safely trying to get private namespace functions}
\usage{
safe_private_export(pkg, f)
}
\arguments{
\item{pkg}{package name as character}

\item{f}{package function name as character}
}
\description{
Predominately, this is used to get private exports from ggplot2 for ggpackets
functions. This behavior is more responsibly handled in ggpackets
(https://www.github.com/dgkf/ggpackets). Once published, gClinBiomarker will
instead include ggpackets as a dependency and this behavior will not be
needed.
}
