% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas-convenience-funcs.R
\name{sas.emmeans}
\alias{sas.emmeans}
\title{Compute emmeans to mirror SAS emmeans functionality}
\usage{
sas.emmeans(model.obj, specs, data = NULL, mode = "kenward-roger",
  quietly = FALSE, verbose = FALSE, confidence.level = NULL, ...)
}
\arguments{
\item{model.obj}{A model object on which to calculate emmeans}

\item{specs}{specs to be passed to emmeans()}

\item{data}{specified data to be used for model fitting. (defaults to data
bound to model, though this can not always been inferred from the model
object and scope of function call)}

\item{mode}{mode to be passed to emmeans()}

\item{quietly}{whether to hide message informing of changes to default
functionality}

\item{verbose}{whether all debug information should be printed to console}

\item{confidence.level}{the confidence interval to use for upper and lower
confidence level outputs}

\item{...}{additional arguments to be passed to emmeans}
}
\description{
Compute emmeans to mirror SAS emmeans functionality
}
\examples{
library(dplyr) # for \%>\% operator
library(nlme)  # for gls

# create model on which to calculate emmeans
model <- gls(mpg ~ hp + carb + wt,
  data = mtcars, na.action = na.exclude)

# call emmeans, specifying factors over which means should be calculated
sas.emm <- sas.emmeans(model, ~ carb, data = mtcars, confidence.level = 0.9)

# update the confidence level used
# NOTE: if you're examining confidence intervals, it's much faster to udpate
# rather than re-run the model and means
sas.emm <- update(sas.emm, level = 0.95)

# using the means as a dataframe
as.data.frame(sas.emm)

}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}, Christina Rabe
  \email{rabe.christina@gene.com}
}
