% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas-convenience-funcs.R
\name{sas.lsmeans}
\alias{sas.lsmeans}
\title{Compute LSMEANS to mirror SAS functionality}
\usage{
sas.lsmeans(model.obj, specs, data = NULL, mode = "kenward-roger",
  quietly = FALSE, verbose = FALSE, confidence.level = NULL, ...)
}
\arguments{
\item{model.obj}{A model object on which to calculate lsmeans}

\item{specs}{specs to be passed to lsmeans()}

\item{mode}{mode to be passed to lsmeans()}

\item{quietly}{whether to hide message informing of changes to default
functionality}

\item{...}{additional arguments to be passed to lsmeans}

\item{as.data.frame}{whether to return a lsm.list object or a data.frame}
}
\description{
Compute LSMEANS to mirror SAS functionality
}
\examples{
library(dplyr) # for \%>\% operator
library(nlme)  # for gls

# create model on which to calculate lsmeans
model <- gls(mpg ~ hp + carb + wt,
  data = mtcars, na.action = na.exclude)

# call lsmeans, specifying factors over which means should be calculated
sas.lsm <- sas.lsmeans(model, ~ carb, data = mtcars, confidence.level = 0.9, verb = T)

# update the confidence level used
# NOTE: if you're examining confidence intervals, it's much faster to udpate
# rather than re-run the model and means
sas.lsm <- update(sas.lsm, level = 0.95)

# using the means as a dataframe
as.data.frame(sas.lsm)

}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}, Christina Rabe
  \email{rabe.christina@gene.com}
}
