% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\docType{methods}
\name{show,ggpacket-method}
\alias{show,ggpacket-method}
\title{Overload show method to print ggpacket}
\usage{
\S4method{show}{ggpacket}(object)
}
\arguments{
\item{object}{the ggpacket object to show}
}
\description{
Overload show method to print ggpacket
}
