% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpkg-.r
\docType{methods}
\name{show,ggpkg-method}
\alias{show,ggpkg-method}
\title{Overload show method to print ggpkg}
\usage{
\S4method{show}{ggpkg}(object)
}
\description{
Overload show method to print ggpkg
}
