% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggwrapper-.r
\docType{methods}
\name{show,ggwrapper-method}
\alias{show,ggwrapper-method}
\title{Overload show method to print ggwrapper}
\usage{
\S4method{show}{ggwrapper}(object)
}
\description{
Overload show method to print ggwrapper
}

