% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-io.r
\name{sink_to_temp}
\alias{sink_to_temp}
\title{Helper function to hide side effect output to console}
\usage{
sink_to_temp(expr)
}
\arguments{
\item{expr}{the expression to be evaluated}
}
\value{
the result of the expression.
}
\description{
Some packages make it very difficult to hide console output of
  their functions. In situations where you want to use those functions
  without cluttering your console output, this function can be used to wrap
  code for evaluation, suppressing all outputs.
}
