% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\name{split_on_name}
\alias{split_on_name}
\title{Helper function to split a list on a named args}
\usage{
split_on_name(l, split_name, omit.named = TRUE)
}
\arguments{
\item{l}{list to split}

\item{split_name}{name of element in list to split on. If the element is not
found in the list, the returned list will contain the entirety of the passed
list, l, in the name 'left' and an empty list in the name 'right'.}

\item{omit.named}{logical indicating whether the list value specified by
split_name should be included in the output lists.}
}
\value{
A named list with two named values, "left" and "right", indicating
arguments to the left and right of the split_name argument.
}
\description{
Specifically used for splitting a call's arguments based on the location
of a specified argument name to use the placement of the arguments to
represent useful information.
}
