% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-stat-summary-funs-.r
\name{stat_summary_funs}
\alias{stat_summary_funs}
\title{helper function to interpret strings and functions passed to
geom_stat_ribbons}
\usage{
stat_summary_funs(format, args)
}
\arguments{
\item{format}{either a function returning a named list of stat
transformations or a string matching to one of 'mean_se', 'median_iqr',
'tukey_hinges', 'tukey_whiskers' or 'quantile_contour', or a list of
strings from this group. The y calculation of the first function in the
list returned is used for placing labels.}

\item{args}{additional arguments to pass to target function}
}
\value{
a list of functions to apply to pass to fun.data in stat_summary
}
\description{
helper function to interpret strings and functions passed to
geom_stat_ribbons
}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}
}
