% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\docType{methods}
\name{[,ggpacket,ANY,ANY,ANY-method}
\alias{[,ggpacket,ANY,ANY,ANY-method}
\title{Overload [ generic to index into ggcalls list}
\usage{
\S4method{[}{ggpacket,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{ggpacket object}

\item{i}{index of the ggcall in the ggpacket object}
}
\description{
Overload [ generic to index into ggcalls list
}
