% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-handling.r
\name{summarize_unary_vars}
\alias{summarize_unary_vars}
\title{Produce summary table of unary value of specified levels}
\usage{
summarize_unary_vars(data, ..., .dots = c())
}
\arguments{
\item{data}{the dataframe to summarized}

\item{...}{character vectors or strings containing variable names to group
over}

\item{.dots}{a character vector ccontaining variables names to group over}
}
\value{
a dataframe with a single value per group combination with only
  variables with a unique value per group.
}
\description{
Given a dataframe and set of grouping variables, a dataset will
  be returned with all groups summarized containing only variables with one
  unique value per group, with a value of that single unique value.
}
