% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumSingle.R
\name{SumSingle}
\alias{SumSingle}
\title{Summary a single covariate, test across treatment and/or population}
\usage{
SumSingle(data, var, trt = NULL, trt.name = NULL, bep = NULL,
  bep.name = NULL, bep.indicator = 1, compare.itt = TRUE,
  itt.name = "ITT", var.class, ordered.factor.levels = NULL,
  cont.show = c("N", "Mean", "Median", "Min-Max", "NA's"), digits = 2,
  trt.order = NULL, test.bep = FALSE, na.action = "error")
}
\arguments{
\item{var}{name of the clinical covariate to test}

\item{trt}{name of the treatment column. If trt is specified, the analysis will be performed within treatment arm.
if it is NULL, the comparison will be performed using all samples.}

\item{trt.name}{preferred display name of the treatment variable
If it is NULL, trt will be used.}

\item{bep}{name of the column which indicates subpopulation (e.g. biomarker evaluable population)}

\item{bep.name}{preferred display name of the subpopulation (e.g. biomarker evaluable population).
If it is NULL, bep will be used.}

\item{bep.indicator}{In the subpopulation column, which value is used
to define the subpopulation (e.g. biomarker evaluable population). 
Default is 1. The non-subpopulation enrties is not allowed to be specified as NA.}

\item{compare.itt}{whether show summary statistics of BEP and ITT. If it is FALSE,
the output will show summary statistics of BEP and nonBEP. Default is TRUE.}

\item{itt.name}{preferred display name of ITT
If it is NULL, "ITT" will be used.}

\item{var.class}{class of the variable. possible categories are "numeric", "categorical" and
"ordered.factor".  "ordered.factor" can be used to categorical variable with
ordered levels - e.g. IC score 0/1/2/3. If class is ordered.factor ,
ordered.factor.levels need to be specified.}

\item{ordered.factor.levels}{ordered levels of the ordered factor.}

\item{cont.show}{what summary statistics to show for a continuous covariate.
Default is c("N" ,"Mean","Median", "Min-Max","NA's").
Possible options are "N" ,"Mean","SEM", "SD","Median",
"Min","Max" ,"Min-Max","1st Qrtl.","3rd Qrtl.","IQR" ,"NA's"}

\item{digits}{digits for rounding}

\item{trt.order}{If the user wants to display the treatments in a certain
order, it can be defined here. All elements in trt.order should be the same
unique values in the treatment column.}

\item{test.bep}{whether test across subpopulations within treatment arm. If class is numeric,
kruskal wallis rank sum test will be performed. If class is categorical, fisher's exact test will be performed.
If class is ordered.factor, cmh test will be performed. The test is always performed between BEP vs nonBEP.
P value columns will be included in the output table if it is specified as TRUE.
Testing is not recommendated if either BEP of non-BEP has small sample size.}

\item{na.action}{defaults to "na.omit". Possible options are "na.omit", "error"
When it is specified as "na.omit", entries with missing trt or bep
will be automatically removed before calculation.}
}
\note{
trt allows for more than 2 levels. However, only 2 levels are allowed for bep.
For more general use, a user can specify trt to get summary statistics for any
sub-group defination (and leave bep as NULL).
}
\author{
Ning Leng, Alexey Pronin, and previous team members (see DESCRIPTION)
}

