% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareKM.R
\name{CompareKM}
\alias{CompareKM}
\title{Compare KM curve of ITT vs BEP}
\usage{
CompareKM(data, tte, cens, trt = NULL, bep, bep.indicator = 1,
  bep.name = "Biomarker Evaluable", itt.name = "ITT",
  col.itt = "palegreen4", col.bep = "lightpink2", col.ci = "lightcyan",
  shaded.ci = TRUE, xlim = NULL, xat = NULL, ylab = paste(tte,
  "Survival Probability"), xlab = "Time", main = "", pdf.name = NULL,
  pdf.param = list(height = 5), par.param = list(mar = c(4, 4, 3, 2)), ...)
}
\arguments{
\item{data}{Input data frame. Rows are patients and columns are variables (e.g. demographics variables, time to event variables, 
biomarker variables, treatment indicator, etc.). One patient per row.}

\item{tte}{column name that indicates the time to event variable}

\item{cens}{column name that indicates the censoring variable associated to tte}

\item{trt}{name of the treatment column. If trt is specified, the analysis will be performed within treatment arm.
if it is NULL, the comparison will be performed using all samples.}

\item{bep}{name of the column which indicates subpopulation (e.g. biomarker evaluable population)}

\item{bep.indicator}{In the subpopulation column, which value is used
to define the subpopulation (e.g. biomarker evaluable population). 
Default is 1. The non-subpopulation enrties is not allowed to be specified as NA.}

\item{bep.name}{preferred display name of the subpopulation (e.g. biomarker evaluable population).
If it is NULL, bep will be used.}

\item{itt.name}{preferred display name of ITT
If it is NULL, "ITT" will be used.}

\item{col.itt, col.bep, col.ci}{Color for itt curve, bep curve and confidence interval curve (CI)}

\item{shaded.ci}{Whether add background shade to disply CI}

\item{xlim, xat, ylab, xlab, main}{see \code{\link{plot}}}

\item{pdf.name}{name of output pdf file. If it is NULL (default), the plots will be displayed but not saved as pdf}

\item{pdf.param}{A list of parameters that define pdf graphics device. See \code{\link{pdf}}. 
for example, pdf.param=list(width=10, height=5)}

\item{par.param}{A list of parameters that define graphcial parameters. See \code{\link{par}}.
for example, par.param=list(mfrow=c(3,1), mar=c(4,4,3,2))}

\item{...}{additional parameters for \code{\link{plot}}}
}
\description{
Need some description since it complains!!!
}
\note{
This function generates KM curves to compare ITT vs BEP, within each treatment arm.
}
\author{
Ning Leng \email{leng.ning@gene.com}, Alexey Pronin \email{pronin.alexey@gene.com}, and previous team members (see DESCRIPTION)
}

