% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\name{flatten_aesthetics_to_group}
\alias{flatten_aesthetics_to_group}
\title{Helper function to flatten specific aesthetics to group}
\usage{
flatten_aesthetics_to_group(mapping, ...)
}
\arguments{
\item{mapping}{a ggplot2 aesthetic mapping}

\item{geom}{a ggplot2 Geom object}
}
\value{
the mapping filtered by accepted aesthetics for the given Geom
}
\description{
Helper function to flatten specific aesthetics to group
}
