% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-ggpackets-.r
\name{ggpack}
\alias{ggpack}
\title{A helper function to wrap ggplot calls, allowing for passing of a
shared list of arguments prefixed by a string. For example,
a call to a custom ggplot template may allow for arguments
'line.color' and 'bar.color' to specify the line and bar colors
seperately. These arguments are parsed appropriately and
passed to the appropriate sub-function.}
\usage{
ggpack(`_call`, args_prefix = NULL, passed_args = NULL, ...,
  null.empty = FALSE)
}
\arguments{
\item{args_prefix}{the prefix string to subset arguments by}

\item{passed_args}{arguments to subset}

\item{...}{additional arguments to use when calling the provided
function}

\item{null.empty}{return NULL if no arguments are received}

\item{_geom}{the function to call}
}
\value{
a call to the specified function with arguments subset for
only those which match the specified prefix
}
\description{
A helper function to wrap ggplot calls, allowing for passing of a
shared list of arguments prefixed by a string. For example,
a call to a custom ggplot template may allow for arguments
'line.color' and 'bar.color' to specify the line and bar colors
seperately. These arguments are parsed appropriately and
passed to the appropriate sub-function.
}
\note{
This function should eventually be split out into a separate
package for wrapping ggplot functions
}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}
}
