% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDens.R
\name{PlotDens}
\alias{PlotDens}
\title{Plot Density}
\usage{
PlotDens(data, biomarker.var, log2 = FALSE, col = "blue", text.font = 3,
  main = paste("Distribution of", biomarker.var), add.num = 0,
  xlab = paste(biomarker.var, ifelse(log2 == TRUE, "(log2 scale)", "")),
  pdf.name = NULL, pdf.param = list(width = 7, height = 7),
  par.param = list(mar = c(4, 4, 3, 2)))
}
\arguments{
\item{data}{Input data frame. Rows are patients and columns are variables (e.g. demographics variables, time to event variables, 
biomarker variables, treatment indicator, etc.). One patient per row.}

\item{biomarker.var}{Name of the variable (e.g. a biomarker variable). Should be in colnames of \code{data}.}

\item{log2}{If TRUE, computes binary (i.e. base 2) logarithm. Default is FALSE.}

\item{col}{The color of the line segments. Default is blue.}

\item{text.font}{Legend text font size. Default is 3.}

\item{main}{The main title. Default is \code{paste("Distribution of", biomarker.var)}.}

\item{add.num}{The constant to add to all values. Helps to avoid applying log transformation on 0 or negative values. Default is 0.}

\item{xlab}{X axis label. Default is \code{paste(biomarker.var, ifelse(log2 == TRUE, "(log2 scale)", ""))}.}

\item{pdf.name}{Name of output pdf file. If it's NULL (default), the plots will be displayed but not saved as pdf.}

\item{pdf.param}{A list of parameters that define pdf graphics device. See \code{\link{pdf}}. Default is \code{list(width=7, height=7)}.}

\item{par.param}{A list of parameters that define graphcial parameters. See \code{\link{par}}. Default is \code{list(mar=c(4,4,3,2))}.}
}
\value{
Histogram and density will be shown. Summary statistics will be also shown on the plots.
}
\description{
Generate density plot for a continuous covariate.
}
\examples{
data(input)
PlotDens(data=input, biomarker.var="OS")
PlotDens(data=input, biomarker.var="PFS", col="red", pdf.name="hist.pdf")
PlotDens(data=input, biomarker.var="PFS", log2=TRUE, pdf.name="hist_log2.pdf")

}
\author{
Alexey Pronin \email{pronin.alexey@gene.com}, Ning Leng \email{leng.ning@gene.com}, and previous team members (see DESCRIPTION)
}

