% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLong.r
\name{PlotLong}
\alias{PlotLong}
\title{Wrapper for fitting model and adjusting to predicted values and
passing newly fitted model to ggplot geom_stat_ribbon}
\usage{
PlotLong(data, mapping, formula = NULL, model = lm, model.args = NULL,
  model.per = NULL, facet.fun = NULL, ...)
}
\arguments{
\item{data}{data to use for model fitting and plotting}

\item{mapping}{ggplot aesthetic mapping to use for plotting}

\item{formula}{formula to use for adjusting values}

\item{model}{model function to use for fitting; defaults to lm}

\item{model.args}{additional model parameters to pass to model}

\item{facet.fun}{function to use for ggplot faceting in ggplot2::facet_grid}

\item{...}{additional arguments passed to underlying ggplot components
by prefixing value with ggplot call name. ("ribbons", "line", "text", "xlab", 
"ylab" or "labs" - e.g. `ribbons.color = 'red'`)}
}
\value{
a ggplot object with y values adjusted based on model function if provided
}
\description{
Wrapper for fitting model and adjusting to predicted values and
passing newly fitted model to ggplot geom_stat_ribbon
}
\examples{
library(tidyverse) # for pipe and tibble
library(broom)     # for augment

# default representation
PlotLong(nasa \%>\% as_tibble, aes(x=month, y=temperature))

# a more appropriate representation for large n
PlotLong(nasa \%>\% as_tibble, aes(x=month, y=temperature), 
         fun.data = 'quartiles')

# include linear adjustment accounting for amount of ozone
PlotLong(nasa \%>\% as_tibble, aes(x=month, y=temperature), 
         formula = temperature ~ ozone, fun.data = 'deciles', 
         show.counts = T)

# adjusting by independent models for the northern and southern hemispheres
PlotLong(nasa \%>\% as_tibble \%>\% mutate(hemi=ifelse(lat>0, "North", "South")), 
         aes(x=month, y=temperature), formula = temperature ~ ozone, 
         model.per = ~ hemi, facet.fun = ~ hemi, fun.data = 'deciles', 
         xlab = "Month", ylab = "Temperature Adjusted for Ozone",
         labs.title = "Temperature by Hemisphere",
         labs.caption = "*idependent models fit per hemisphere")

}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}
}
