% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotParam.R
\name{PlotParam}
\alias{PlotParam}
\title{Set graphical parameters}
\usage{
PlotParam(pdf.name, pdf.param, par.param)
}
\arguments{
\item{pdf.name}{A name of a pdf file.}

\item{pdf.param}{A list of parameters that define pdf graphics device. See \code{\link{pdf}}.}

\item{par.param}{A list of parameters that define graphcial parameters. See \code{\link{par}}.}
}
\description{
\code{PlotParam} can be used to set graphical parameters for a pdf file
or the default screen device.
}
\note{
The \code{PlotParam} function is designed to restore graphical parameters \code{par}
to the default values after execution.
}
\examples{
pdf.param = list(height=5)
par.param = list(mar=c(4, 4, 3, 2))
pdf.name = NULL
PlotParam(pdf.name, pdf.param, par.param)
hist(rnorm(1000), col="blue")
PlotParam()

}
\author{
Alexey Pronin \email{pronin.alexey@gene.com}, Ning Leng \email{leng.ning@gene.com}
}
