% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSTEPP.R
\name{PlotSTEPP}
\alias{PlotSTEPP}
\title{Create a STEPP (Subpopulation Treatment Effect Pattern Plot)}
\usage{
PlotSTEPP(data, outcome.var, outcome.class, treatment.var, biomarker.var,
  covariate.var = NULL, strat.var = NULL, placebo.code, active.code,
  alpha = 0.05, window.width = 0.25, min.pt = NULL, max.pt = NULL,
  by = 0.05, yrange.lower = NULL, yrange.upper = NULL, xticks = NULL,
  show.refline = TRUE, refline.color = "grey", show.refline.ac = TRUE,
  refline.color.ac = "lightblue", estimate.color = "blue",
  estimate.lty = 1, estimate.lwd = 2, ci.color = "black", ci.lty = 2,
  ci.lwd = 1, ci.shade = TRUE,
  plot.title = "STEPP: Subgroup Treatment Effect Pattern Plot",
  sub.title = NULL, xlabel = "Biomarker Percentile", ylabel = NULL,
  show.legend = TRUE, legend.loc = "topright",
  legend.text = c("Point Estimate", paste((1 - alpha) * 100,
  " percent Confidence Interval", sep = "")), legend.col = c(estimate.color,
  ci.color), legend.lty = c(estimate.lty, ci.lty),
  legend.lwd = c(estimate.lwd, ci.lwd), legend.bty = "n",
  pdf.name = paste("STEPP ", as.character(Sys.Date()), ".pdf", sep = ""),
  pdf.param = list(width = 11, height = 8.5), par.param = list(mar = c(4, 4,
  3, 2)), csv.name = paste("STEPP ", as.character(Sys.Date()), ".csv", sep =
  ""), bm.digits = 2)
}
\arguments{
\item{data}{input data frame. Rows are patients and columns are variables (e.g. demographics variables, time to event variables,
biomarker variables, treatment indicator, etc.). One patient per row.}

\item{outcome.var}{outcome variable. In case of a 'survival', variable, it will be a vector of two variables: 1) time to event 2) censorship}

\item{outcome.class}{outcome class of the 'outcome' variable. Can be either "continuous", "binary", or "survival".}

\item{treatment.var}{name of the treatment variable.}

\item{biomarker.var}{name of the biomarker variable.}

\item{covariate.var}{vector specifying the covariate variables.
This can be added to adjust for in the analysis for survival and continuous outcome variable classes.}

\item{strat.var}{vector specifying the stratification variables. This can be added for the survival outcome variable class.}

\item{placebo.code}{name of the control group within the treatment variable}

\item{active.code}{name of the treatment/experimental group within the treatment variable}

\item{alpha}{confidence level (CI) for point estimate, i.e. 0.05 for 95 percent CI. Default is 0.05.}

\item{window.width}{width of each window. Default is 0.25.}

\item{min.pt}{minimum center.pt. Default is NULL.}

\item{max.pt}{maximum center.pt. Default is NULL.}

\item{by}{size of 'slide', i.e. speed of the window that moves along the x-axis. Default is 0.05.}

\item{yrange.lower}{value of the lower y-axis. Default is NULL.}

\item{yrange.upper}{value of the upper y-axis. Default is NULL.}

\item{xticks}{x-tick marks. Default is NULL.}

\item{show.refline}{if TRUE, the reference line will be displayed. Default is TRUE.}

\item{refline.color}{color of the reference line. Default is "grey".}

\item{show.refline.ac}{if TRUE, the reference line for effect in All Comers will be displayed. Default is TRUE.}

\item{refline.color.ac}{color of the reference line for effect in All Comers. Default is "lightblue".}

\item{estimate.color}{color of the estimate line. Default is "blue".}

\item{estimate.lty}{type of the estimate line. Default is 1.}

\item{estimate.lwd}{width of the estimate line. Default is 2.}

\item{ci.color}{color of the CI lines. Default is "black".}

\item{ci.lty}{type of the CI lines. Default is 2.}

\item{ci.lwd}{width of the CI lines. Default is 1.}

\item{ci.shade}{if TRUE, the area between the lower and upper CI lines will be shaded in gray. Default is TRUE.}

\item{plot.title}{title of the plot. Default is "STEPP: Subgroup Treatment Effect Pattern Plot".}

\item{sub.title}{subtitle of the plot, this will be displayed on the bottom of the plot. Default is NULL.}

\item{xlabel}{x-axis label. Default is "Biomarker Percentile".}

\item{ylabel}{y-axis label. Default is NULL.}

\item{show.legend}{if TRUE, a legend will be displayed. Default is TRUE.}

\item{legend.loc}{location of the legend to be displayed. Default is "topright".}

\item{legend.text}{text to be displayed in the legend.
Default is \code{c("Point Estimate",paste((1-alpha)*100," percent Confidence Interval",sep=""))}}

\item{legend.col}{colors of the legend. Default is \code{c(estimate.color, ci.color)}.}

\item{legend.lty}{type of the legend lines. Default is \code{c(estimate.lty, ci.lty)}.}

\item{legend.lwd}{width of the legend lines. Default is \code{c(estimate.lwd, ci.lwd)}.}

\item{legend.bty}{type of box for the legend. Default is "n".}

\item{pdf.name}{name of output pdf file. If it's NULL, the plots will be displayed but not saved as pdf.
Default is \code{paste("STEPP ", as.character(Sys.Date()), ".pdf", sep = "")}.}

\item{pdf.param}{a list of parameters that define pdf graphics device. See \code{\link{pdf}}. Default is \code{list(width=11, height=8.5)}.}

\item{par.param}{a list of parameters that define graphcial parameters. See \code{\link{par}}. Default is \code{list(mar=c(4,4,3,2))}.}

\item{csv.name}{csv file name (includes numbers used in the graphs). If NULL, the function will return the result.
Default is \code{paste("STEPP ", as.character(Sys.Date()), ".csv", sep = "")}.}

\item{bm.digits}{digits to be displayed/used for the lower and upper confidence level estimates. Default is 2.}
}
\description{
This function creates a STEPP from the given point estimates and confidence intervals at desired percentiles.
}
\note{
Patient data without corresponding biomarker data are automatically removed.
For survival data, censorship variable is 1 if an event happened, 0 if censored.
}
\examples{

data(input)
PlotSTEPP(data = input,
         outcome.var = c("PFS", "PFS.CNSR"),
         outcome.class = "survival",
         treatment.var = "Arm",
         biomarker.var = "KRAS.exprs",
         covariate.var = "Sex",
         strat.var = "Age",
         placebo.code = "CTRL",
         active.code = "TRT",
         csv.name = NULL,
         pdf.name = NULL)

}
\author{
Alexey Pronin \email{pronin.alexey@gene.com}, Ning Leng \email{leng.ning@gene.com}, and previous team members (see DESCRIPTION)
}
\seealso{
\link{StatSummary}
}

