% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTabForestMulti.R
\name{PlotTabForestMulti}
\alias{PlotTabForestMulti}
\title{Generate forest plot and summarization table for 2-arm or within-arm comparison for multiple variables}
\usage{
PlotTabForestMulti(data, outcome.class = c("survival", "binary"), outcome.var,
  trt = NULL, var, var.class = NULL, var.name = NULL,
  percentile.cutoff = 0.5, greater = TRUE, less = TRUE,
  within.bin = FALSE, compare.bep.itt = TRUE, compare.subgroup = FALSE,
  show.itt = FALSE, show.bep = FALSE, subgroup = NULL, bep = NULL,
  bep.name = "BEP", itt.name = "All", bep.indicator = 1,
  covariate = NULL, strata = NULL, quantile.type = 2,
  placebo.code = NULL, active.code = NULL, var.code = NULL,
  tabforest = FALSE, alpha = 0.05, main = NULL, sub = NULL,
  clip = NULL, xlab = NULL, cex.headings = 1.1, cex.note = 1,
  cols = "darkgreen", pdf.name = NULL, pdf.param = list(width = 6, height
  = 4.5), par.param = list(cex = 1.2, cex.main = 1.5, cex.sub = 1, cex.axis =
  1))
}
\arguments{
\item{data}{input data frame. Rows are patients and columns are variables (e.g. demographics variables, time to event variables,
biomarker variables, treatment indicator, etc.). One patient per row.}

\item{outcome.class}{type of the outcome variable. Default is \code{c("survival", "binary", "continuous")}}

\item{outcome.var}{name of the outcome varible. If the outcome.class is binary or coutinuous, only one value should be provided.
If the outcome.class is survival, two values should be provided - name of the 'time to event' variable and 'censorship' variable
 For the censoring variable, 1 indicates event and 0 indicates censoring}

\item{trt}{name of the treatment variable. If this is NULL, within-arm analysis will be performed}

\item{var}{name of the biomarker variable. only one variable should be specified.}

\item{var.class}{class of the variable. valid categories are "numeric", "categorical". If the class is continuous,
user needs to specify percentile.cutoff to dichotomize the continuous measure into subgroups}

\item{var.name}{display name for the biomarker variable}

\item{percentile.cutoff}{percentile to dichotomize continuous biomarker measure. This could be a vector with multiple elements.
Values should be between 0 and 1}

\item{greater}{whether calculate summary statistics within the subgroup whose biomarker value is greater than or equal to
cutoff value. If this is TRUE, in 2-arm study, across-arm HR within biomarker high group will be calculated.
In single arm study HR of biomarker high vs low will be calculated.}

\item{less}{whether calculate summary statistics within the subgroup whose biomarker value is less than the cutoff value.
greater and less can both be TRUE. If compare across subgroups (compare.subgroup=TRUE), both greater and less will be set as TRUE}

\item{within.bin}{whether calculate summary statistics within bin (e.g. > cutoff1 and <= cutoff2). If within.bin is TRUE,
greater and less will be set as FALSE.}

\item{compare.bep.itt}{whether want to generate two groups of results to compare the summary statistics
in ITT vs in BEP. If this is TRUE, paramemeter bep should be specified. If this is FALSE,
parameters bep, bep.name, itt.name, bep.indicator will be ignored}

\item{compare.subgroup}{whether want to generate multiple groups of results to compare the summary statistics
in each subgroup defined by parameter subgroup}

\item{show.itt}{when performing subgroup comparison (compare.subgroup=TRUE), whether also calculate summary statistics using all patients in itt}

\item{show.bep}{when performing subgroup comparison (compare.subgroup=TRUE), whether also calculate summary statistics using all patients in BEP (biomarker evaluable population).
BEP is defined by variable bep}

\item{subgroup}{The column which defines subgroups. If compare.subgroup is TRUE, the program will generate forest plot of the vars within each subgroup}

\item{bep}{name of the column which indicates biomarker evaluable population.}

\item{bep.name}{preferred display name of the biomarker evaluable population.
If it is NULL, bep will be used.}

\item{itt.name}{preferred display name of full population
If it is NULL, "All" will be used.}

\item{bep.indicator}{In the subpopulation column, which value is used
to define the biomarker evaluable population.}

\item{covariate}{a vector specifying the covariate variables to be adjusted in the model. Default is set to NULL, meaning no adjustment.}

\item{strata}{name of the stratification variables. Default is set to NULL, meaning no stratification.}

\item{placebo.code}{name of the control arm of the treatment variable}

\item{active.code}{of the treatment/experimental arm of the treatment variable}

\item{var.code}{ordered levels of the biomarker variable. This will be ignored for continuous biomarker.
If the biomarker is categorical and this is NULL, biomarker subgroups will be ordered by the order from factor() function}

\item{tabforest}{Default is FALSE. If it is FALSE, forest plot will be generated using forestplot() function.
If it is TRUE, a table will be generated with forest plots incorpriated}

\item{alpha}{type I error rate. Default is 0.05.}

\item{main}{main title of the forest plot. Default is "Association of biomarker effect within treatment arms".}

\item{sub}{sub title under the forest plot. Default is NULL.}

\item{clip}{range of the x-axis of the forest plot. Default is NULL.}

\item{cex.headings}{amount of magnification of headings of the forest plot relative to cex. Default is 1.1.}

\item{cex.note}{amount of magnification of the note. Default is 1.}

\item{cols}{Color of the 'effect size' displayed in the forest plot.}

\item{pdf.name}{name of output pdf file. If it's NULL, the plots will be displayed but not saved as pdf. Default is "Forestplot.pdf".}

\item{pdf.param}{a list of parameters that define pdf graphics device. See \code{\link{pdf}}. Default is \code{list(width=6, height=4.5)}.}

\item{par.param}{a list of parameters that define graphcial parameters. See \code{\link{par}}. Default is \code{list(mar=c(4,4,3,2))}.}
}
\description{
This function creates a forest plot along with table with summary statistics to infer effect of multiple clinical variables, within a single arm
or across two treatment arms.  The outcome could be survival, binary or continuous. This function can be used to summarize a group of
variables. This function may be used to compare effect of these variables in full population vs biomarker evaluable population. Or compare effect
of these variables in different biomarker subgroups
}
\examples{
data(input)
PlotTabForestBiomarker(data=input,
                      outcome.class=c("survival"),
                      outcome.var=c("PFS","PFS.event"),
                      trt="Arm",
                      var="KRAS.mutant",
                      var.class="categorical")
}
\author{
Ning Leng \email{leng.ning@gene.com}, Alexey Pronin \email{pronin.alexey@gene.com}, and previous team members (see DESCRIPTION)
}

