% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{ReadData}
\alias{ReadData}
\title{Read data}
\usage{
ReadData(input, input.specs = NULL)
}
\arguments{
\item{input}{A csv file with data.}

\item{input.specs}{A csv file with specs for \code{input}. Default is NULL.}
}
\value{
A list with two data frames: data and data.specs.
}
\description{
Read two csv files and perform concordance check. 
First file contains data.
Second file provides specs for data.
}
\note{
The ReadData() function reads in the data file and spec file from csv files. 
The function also perform concordance check between the data and its spec. 
Data column names and spec entry names should match. No duplicate names are allowed.
If input.specs is NULL, it will be created based on the input data and then returned without checks.
}
\examples{
\dontrun{
ReadData("input.csv", "input_specs.csv")
ReadData("input.csv")
}

}
\author{
Alexey Pronin \email{pronin.alexey@gene.com}, Ning Leng \email{leng.ning@gene.com}
}
