% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadData}
\alias{ReadData}
\title{Read data}
\usage{
ReadData(input, input.specs)
}
\arguments{
\item{input}{A file with data.}

\item{input.specs}{A file with specs for \code{input}.}
}
\value{
Two data frames: df and df.specs
}
\description{
Read two files. First file contains data.
Second file provides specs for data.
}
\examples{
\dontrun{
ReadData("input.csv", "input_specs.csv")
}

}
\author{
Alexey Pronin
}

