% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vad.R
\name{ReadVAD}
\alias{ReadVAD}
\title{Read a VAD}
\usage{
ReadVAD(file, data.write.to = NULL, data.specs.write.to = NULL)
}
\arguments{
\item{file}{A SAS dataset.}

\item{data.write.to}{A csv file where the data will be written to. Default is NULL.}

\item{data.specs.write.to}{A csv file where the data specs will be written to. Default is NULL.}
}
\value{
A list with two data frames: data and data.specs.
}
\description{
Read a VAD and create specs for it.
}
\note{
The ReadVAD() function reads in a VAD and creates data specs.
}
\examples{
\dontrun{
ReadVAD("ars.sas7bdat", data.write.to = "data.csv", data.specs.write.to = "data.specs.csv")
ReadVAD("ars.sas7bdat")
}

}
\author{
Alexey Pronin
}

