% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas-convenience-funcs.R
\name{sas.lsmeans}
\alias{sas.lsmeans}
\title{Compute LSMEANS to mirror SAS functionality}
\usage{
sas.lsmeans(model.obj, specs, mode = "kenward-roger", as.data.frame = FALSE,
  quietly = FALSE, ...)
}
\arguments{
\item{model.obj}{A model object on which to calculate lsmeans}

\item{specs}{specs to be passed to lsmeans()}

\item{mode}{mode to be passed to lsmeans()}

\item{as.data.frame}{whether to return a lsm.list object or a data.frame}

\item{quietly}{whether to hide message informing of changes to default functionality}

\item{...}{additional arguments to be passed to lsmeans}
}
\description{
Compute LSMEANS to mirror SAS functionality
}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}, Christina Rabe \email{rabe.christina@gene.com}
}
