% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas-convenience-funcs.R
\name{sas.lsmeans}
\alias{sas.lsmeans}
\title{Compute LSMEANS to mirror SAS functionality}
\usage{
sas.lsmeans(model.obj, specs, data = NULL, mode = "kenward-roger",
  quietly = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{model.obj}{A model object on which to calculate lsmeans}

\item{specs}{specs to be passed to lsmeans()}

\item{mode}{mode to be passed to lsmeans()}

\item{quietly}{whether to hide message informing of changes to default functionality}

\item{...}{additional arguments to be passed to lsmeans}

\item{as.data.frame}{whether to return a lsm.list object or a data.frame}
}
\description{
Compute LSMEANS to mirror SAS functionality
}
\examples{
library(dplyr) # for \%>\% operator

# create model on which to calculate lsmeans
gls(mpg ~ hp + carb + wt,
    data = mtcars \%>\%
             mutate(carb = as.factor(carb)), # make sure contrast is factor
    na.action = na.exclude) \%>\%

# call lsmeans, specifying factors over which means should be calculated
sas.lsmeans(~ carb)

}
\author{
Doug Kelkhoff \email{kelkhoff.douglas@gene.com}, Christina Rabe \email{rabe.christina@gene.com}
}
